// src/features/config/types.ts
/*
    artifact generator: C:\My\wizzi\stfnbssl\wizzi.plugins\packages\wizzi.plugin.ts\lib\artifacts\ts\module\gen\main.js
    package: @wizzi/plugin.ts@
    primary source IttfDocument: C:\My\wizzi\stfnbssl\wizzi.apps\packages\wizzi.hub.backend\.wizzi-override\src\features\config\types.ts.ittf
    utc time: Fri, 24 Jan 2025 15:11:51 GMT
*/
export type ConfigType = { 
    noCache: boolean;
    sessionSecret: string;
    mongoHost: string;
    mongoUser: string;
    mongoPassword: string;
    mongoPath: string;
    packiConfigPath: string;
    isWizziDev: boolean;
    port: number;
    userUserName: string;
    userDisplayName: string;
    userAvatarUrl: string;
    ittfPath: string;
    dataPath: string;
    metaHtmlIttfPath: string;
    metaFolderIttfPath: string;
    metaHtmlTextPath: string;
};
;